Always dismiss keyboard on search result selection
Issue:

It appears that
https://github.com/rnystrom/GitHawk/commit/66bdf3d835528bba22c8a0ff64e7b8d958178bf0
introduced a bug with search result selection.

After the change from `navigationController?.pushViewController` to
`showDetailViewController`, it was possible to enter a state
where the `RepositoryViewController` would be pushed onto the stack,
 without dismissing the keyboard.

To recreate:

1. Have at least 1 recent search.
2. Select a recent search item
 - search bar will become first responder
3. Tap in the search bar to modify the search text, DO NOT tap "search"
4. Select an item in the list from the original search results.
5. RepositoryViewController will be pushed onto the stack, but the
keyboard will remain visible.

Solution:

Declare and implement `SearchResultSectionControllerDelegate`. This delegate is used
to notify the `SearchViewController` that a search result has been selected.

In turn, the `SearchViewController` will resign first responder status
of its search bar.