switch from the now deprecated to the new statuspage
the "old" status page https://status.github.com is no deprecated and
GitHub switched to a statuspage.io implementation with
https://githubstatus.com. The page mentions the deprecation.

This PR switches the url to the new status page and changes the status
API model for the new REST-response.

The new API replaced the `good` state with the state `none`. And has more
fields, that are probably not relevant for the App.