Edit issue title
This has three pieces of key functionality:
* sends a v3 network request to update the issue title
* updates the IssueTitleSectionController (the title the user sees
when they open an Issue)
* updates the Issue Timeline

These can be tested/verified manually. I also added a unit test
to verify that the User's interaction with the EditIssueTitleViewController
works as expected, eg a network request won't be sent unless a user
edits the issue title and calls onSave().

This is an update of #2348 - in this version of the PR,
 EditIssueTitle's behavior is intended to be very similar to
 Labels, People, and Milestones. Ie it updates the UI optimistically and
 these updates are triggered by ContextMenuWillDismiss().

I didn't add any funcionality to update Bookmarks after an
issue title is edited because my understanding is that the
bookmark titles are not stored anymore but are retrieved from
 the github api.

I updated some string constants (requested in original pr) and I removed the
 width parameter from a helper function [titleStringSizing] because
it was convenient and I couldn't find anywhere that parameter was used

[titleStringSizing]: https://github.com/GitHawkApp/GitHawk/blob/45e9cf6428d5edfe00a73eb18450c50c03871510/Classes/Issues/IssueViewModels.swift#L13