Fix for “Disallow Empty Bug Reports” Issue # 2615
Added a guard to the OnSend function that looks and sees verifies the bodyField.text is not nil, only contains whitespace, or contains no text at all before submitting issue.

Called Squawk to display an error message in the bodyField view container as it had been displaying behind the keyboard.

Lastly, changed the Squawk being used when an empty titleText.text field is detected to also display in the bodyField container as it would also be hidden by the keyboard.