Fix for "Disallow Empty Bug Reports" Issue #2615
Reverted to original code for function titleFieldEditingChanged. Removed textViewDelegate and updateSubmitButtonState function which watched both titleField and bodyField for content before enabling the Submit button. 

Split onSend function in two separate functions, onSend and finishSend. onSend now looks for content in bodyField and if none is found instantiates an IAAlertController to notify the user that they are missing a description and giving them the option to send it anyway or return to the form and add one. finishSend completes the send.